+++ src/main/java/br/com/sis/rh/apiprogramaformacao/api/vo/form/ProcessoSeletivoForm.java
@@ -11,9 +11,9 @@ public class ProcessoSeletivoForm {
     private String nomeInstrutor;
     private LocalDate dataInicio;
     private LocalDate dataFim;
-    private Long qtdEstagiario;
-    private Long qtdTrainees;
-    private Long qtdAprendizes;
+    private Integer qtdEstagiario;
+    private Integer qtdTrainees;
+    private Integer qtdAprendizes;
 
     public ProcessoSeletivo converter(InstrutorRepository instrutorRepository){
         ProcessoSeletivo processoSeletivo = new ProcessoSeletivo();
@@ -61,27 +61,27 @@ public class ProcessoSeletivoForm {
         this.dataFim = dataFim;
     }
 
-    public Long getQtdEstagiario() {
+    public Integer getQtdEstagiario() {
         return qtdEstagiario;
     }
 
-    public void setQtdEstagiario(Long qtdEstagiario) {
+    public void setQtdEstagiario(Integer qtdEstagiario) {
         this.qtdEstagiario = qtdEstagiario;
     }
 
-    public Long getQtdTrainees() {
+    public Integer getQtdTrainees() {
         return qtdTrainees;
     }
 
-    public void setQtdTrainees(Long qtdTrainees) {
+    public void setQtdTrainees(Integer qtdTrainees) {
         this.qtdTrainees = qtdTrainees;
     }
 
-    public Long getQtdAprendizes() {
+    public Integer getQtdAprendizes() {
         return qtdAprendizes;
     }
 
-    public void setQtdAprendizes(Long qtdAprendizes) {
+    public void setQtdAprendizes(Integer qtdAprendizes) {
         this.qtdAprendizes = qtdAprendizes;
     }
 }
